# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrackMovil — API",
    "version": "0.1.0",
    "description": "Catálogo de las APIs propias de TrackMovil (RiogasTracking, app 5 en SecuritySuite).\n\nGENERADO por `pnpm docs:api` a partir de app/api/**/route.ts. No editar a mano:\nlas descripciones, los consumidores y los ejemplos van en docs/api/anotaciones.yaml,\nque se mergea encima al servir GET /api/docs/spec.\n\n`x-auth` refleja los gates que el handler REALMENTE invoca. `x-auth.sinGate: true`\nsignifica que el endpoint no valida nada: es información sensible y por eso este\nportal es solo-root."
  },
  "servers": [
    {
      "url": "https://track.glp.riogas.com.uy",
      "description": "producción"
    }
  ],
...
  "paths": {
...
    "/api/auth/permisos": {
      "post": {
        "operationId": "postApiAuthPermisos",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Request inválido"
          },
          "401": {
            "description": "No autenticado"
          },
          "502": {
            "description": "Error del upstream"
          }
        },
        "summary": "POST /api/auth/permisos",
        "tags": [
          "auth"
        ],
        "x-archivo": "app/api/auth/permisos/route.ts",
        "x-auth": {
          "funcionalidades": [],
          "gates": [
            "requireAuthorizationHeader (token para el SecuritySuite)"
          ],
          "sinGate": false
        },
        "x-generado": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sesionSupabase": {
        "type": "http",
        "scheme": "bearer",
        "description": "Sesión de Supabase Auth validada por requireAuth (lib/auth-middleware.ts)."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key de integraciones externas (sender GeneXus/SGM, VB6)."
      },
      "tokenSecapi": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT emitido por SecuritySuite. Es el que verifica lib/docs/root-guard.ts contra secapi."
      }
    }
  },
...
}